Modifying tuning curve implementation
Speed-up using np.histogram (instead of basically
implementing this ourselves). Including tests for this
new implementation. Input to tuning curves is now binsize,
which is the desired size for each position bin (in whatever
units the position is in, cm or pixels). This is instead of
the num_bins from before that required the calculations to
be computed outside of the function. Edges are based on this
binsize and the length of the linearized position.