Adding test for linear position
The linear position is in terms of the ideal trajectory
line, not in global position points. This is probably
what we want for tuning curves anyways, but it's good to keep
this in mind in case it ever becomes useful to also have the global
position points. They can be gotten from using the interpolate
shapely function on the ideal_path line.