Test and refactoring co-occurrence
Much more readable when this huge function was broken up
into smaller, more specific functions. Functionality is
pretty similar for users with a core function that calls
all these more specific helper functions. Possible to
include the more specific functions in the vdmlab namespace,
but let's wait until we have a use for it to keep things simple.