Updates to decoding bayesian prob likelihood
Strange behavior when both the position and time bins
are large. Infinity raises a warning and the output is
a nan, when really that bin has highest likelihood.

Some refactoring to better understand the underlying math.
From the poisson likelihood, it seems we're not dividing by
the factorial of counts(i). Will look into this.

Strangely, the output of decoding error when using actual data
is the same as before. Will look into what is happening in the
other functions involved.