Decoding 2D
Modified decoding 1D to work as similar as 2D as possible. Mostly,
the differences between them are in the script calling these functions.

This code is relatively untested. It works on simple datasets in
a notebook with the expected results and runs on some sessions from
the shortcut project, but hasn't been tested further than that.