Implementing co-occurrence with 2D tuning curves
In practice, this mostly means that place detection was implemented
in the 2D tuning curves. But other modifications were also done
while this was implemented.

Specifically, sharp-wave ripple events are now vdmlab.Epoch. The
vdmlab.Epoch also got a new property that returns the number of
epochs.

New feature that finds if multiple neurons were active during a given
epoch, and returns a new set of epochs that only include those that
have a min number of active neurons.

Sampling_rate in computing the tuning curves is now computed from the
data instead of being given to the function "hard-coded".