Extending Epochs to have a contains method and adding units to scalebar
The contains method will return an Epoch object with the same start,
stop as self for every epoch that overlaps with epoch_b.

Rounding and adding seconds to scalebar function for more attractive
plotting.

remove_teleports now returns an Epoch object, which can be converted
to position with position_epoch(). This was done because it seemed easier
to compare the no-teleport times to SWR-times if they were both epoch
objects.

Some basic tests for the new features in Epoch and changes to
remove_teleports().