Separate Epoch method intersect and overlaps
Here the method `intersect' only returns the intersection between
two sets of epochs. A new method `overlaps' deals with the case
where there is a template Epoch and the desired behavior is to return
all epochs of interest that overlaps at all with the template epochs.
This means that for edge cases where the epochs of interest start before
or end after the template epoch, the starts and stops from the epoch of
interest are returned and there are therefore some times that do not
intersect between the two sets of epochs.