Remove some bias from bin edges
Previously, the first bin started at the minimum of the position.
The last bin would have some leftover bit, up to binsize. Now,
the leftover is calculated first and the minimum is adjusted to
take half of the leftover.

Also, we return the occupancy along with the tuning curve in
tuning_curve_1d.